refactor(auth): update auth imports to new paths
