+++ client/src/components/pages/Store/Wallet/__tests__/StoreWallet.test.jsx
@@ -4,7 +4,7 @@ import userEvent from "@testing-library/user-event";
 import * as useModulesHook from "@hooks/useModules";
 import * as usePaymentWebsocketHook from "@hooks/usePaymentWebsocket";
 import { I18nProvider } from "@i18n/I18nProvider";
-import { AuthContext } from "@modules/auth/AuthProvider";
+import { AuthContext } from "@/providers/auth/AuthProvider";
 import * as cashierService from "@modules/cashier/cashierService";
 import * as configurationsProvider from "@providers/configurations/configurationsProvider";
